+++ demos/basic/urls.py
@@ -4,7 +4,6 @@ from django.views.static import serve
 from django.conf.urls import include, url
 from django.views.generic import TemplateView, RedirectView
 
-admin.autodiscover()
 doc_root = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'docs/build/html')
 
 urlpatterns = [