+++ swingtime/admin.py
@@ -15,16 +15,11 @@ class OccurrenceInline(admin.TabularInline):
     extra = 1
 
 
-class EventNoteInline(GenericTabularInline):
-    model = Note
-    extra = 1
-
-
 class EventAdmin(admin.ModelAdmin):
     list_display = ('title', 'event_type', 'description')
     list_filter = ('event_type', )
     search_fields = ('title', 'description')
-    inlines = [EventNoteInline, OccurrenceInline]
+    inlines = [OccurrenceInline]
 
 
 admin.site.register(Event, EventAdmin)